+++ titan/common/operators.py
