+++ titan/common/type.py
